Refactor spring
